+++ lib/utils/AbstractQueryConfiguration.dart
@@ -131,5 +131,6 @@ abstract class AbstractQueryConfiguration {
         getParametersMap(),
         user,
         queryType: queryType,
+        addCredentialsToBody: false,
       );
 }